requirements.txt
beautifulsoup4==4.11.1
Flask==1.1.2
google_api_python_client==2.54.0
joblib==1.1.0
matplotlib==3.5.1
numpy==1.21.5
pandas==1.4.2
Pillow==9.2.0
protobuf==4.21.2
pymongo==4.1.1
requests==2.27.1
scikit_learn==1.1.1
seaborn==0.11.2
streamlit==1.10.0
wordcloud==1.8.2.2
xgboost==1.6.1
